fix: define estrutura Item
