+++ src/classes/item/item.hpp
@@ -1,5 +1,5 @@
-#ifndef ITME_H
-#define ITEM_H
+#ifndef ITME1_H
+#define ITEM1_H
 
 #include <iostream>
 #include <string>
